+++ cx-portal/src/features/apps/apiSlice.ts
@@ -40,7 +40,7 @@ export type AppMarketplaceApp = {
   provider: string
   leadPictureUri: string
   shortDescription: string
-  useCases: string[]
+  useCases: UseCase[]
   price: string
   rating?: number
   uri?: string
@@ -50,6 +50,11 @@ export type AppMarketplaceApp = {
   lastChanged?: string
   timestamp?: number
   leadPictureId?: string
+}
+
+export type UseCase = {
+  id: string
+  label: string
 }
 
 export enum SubscriptionStatus {